extract common logging infrastructure.
